backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.10.0
openai==1.51.0
python-dotenv==1.0.1
matplotlib==3.9.2
matplotlib-venn==1.1.1
pydantic==2.9.2
